replace Malte2.Server with Malte2.Service which can be a Windows service
