--- Malte2.Server/.gitignore
@@ -1,1 +1,0 @@
-/*.sqlite
