+++ Malte2.Service/Application/LogConstants.cs
