+++ Malte2.Service/Database/DatabaseUpdater.cs
@@ -232,6 +232,8 @@ ALTER TABLE operation ADD COLUMN invoice TEXT;
             // upgrade tables
             applyCommand(databaseContext, transaction, @"
 
+DROP TRIGGER ck_operation_category_entry;
+
 CREATE TRIGGER ck_operation_category_entry
 BEFORE UPDATE OF category_id, accounting_entry_id ON operation
 FOR EACH ROW WHEN NEW.category_id IS NOT NULL AND NEW.accounting_entry_id NOT IN (SELECT accounting_entry_id FROM accounting_category WHERE accounting_category_id = NEW.category_id)
