+++ Malte2.Service/Database/DatabaseValueUtils.cs
