+++ Malte2.Service/Database/DateTimeDatabaseUtils.cs
