+++ Malte2.Service/Model/Accounting/OperationEditionCsvLine.cs
