+++ Malte2.Service/Model/Accounting/OperationEditionType.cs
