+++ Malte2.Service/ServiceUninstall.bat
@@ -1,0 +1,1 @@
+SC DELETE "Malte2"
