+++ Malte2.Service/Services/ServiceConfiguration.cs
