+++ Malte2.Service/StartMalte2.bat
@@ -1,0 +1,1 @@
+start "" "http://localhost:5000"
