+++ Malte2.Service/appsettings.json
@@ -4,6 +4,11 @@
       "Default": "Information",
       "Microsoft": "Warning",
       "Microsoft.Hosting.Lifetime": "Information"
+    },
+    "EventLog": {
+      "LogLevel": {
+        "Default": "Warning"
+      }
     }
   },
   "AllowedHosts": "localhost;127.0.0.1",
